Increase the NetuitiveHandler batch size to 500
Incremental posts are synchronous calls made during collection if the
number of samples cached is equal to the batch size. These are invasive,
heavy operations which slow down collection and use a lot of CPU. Each
collector flushes data after collection is completed, so increasing the
batch size won't affect the timeliness of data.